+++ src/lib.rs
@@ -1,12 +1,17 @@
 mod cast;
 mod containing_float;
+mod float;
 mod goniometric;
+mod into_float;
 mod isqrt;
+mod one;
 mod sqrt;
 mod vec2;
+mod vec2_range_iterator;
 
 pub use self::{
-    cast::*, containing_float::*, goniometric::*, isqrt::*, sqrt::*, vec2::*,
+    cast::*, containing_float::*, float::*, goniometric::*, into_float::*,
+    isqrt::*, one::*, sqrt::*, vec2::*, vec2_range_iterator::*,
 };
 
 #[cfg(test)]