+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # CHANGELOG
 
 ## future
-- Add new methods for `Vec2`: `pos_of_idx` and `angle`.
-- Add new traits: `ContainingFloat` and `Goniometric`.
+- Add new methods for `Vec2`: `pos_of_idx`, `angle`, `normalized`,
+  `norzmalize`, `to` and `from_polar`.
+- Add new traits: `ContainingFloat`, `Goniometric`, `Float`, `IntoFloat` and
+  `One`.
+- Add `Vec2Range`.
+- Implement `Cast` for float types.
+- Implement `Neg` for `Vec2`.
 
 ## v0.1.0
 - Add `Vec2`.